dev: auto refresh playable list after uploading custom set
